+++ internal/whatsappdb/media_materialized_other.go
@@ -1,0 +1,7 @@
+//go:build !darwin
+
+package whatsappdb
+
+import "os"
+
+func fileMaterialized(os.FileInfo) bool { return true }